+++ tests/testthat/test-oauth.R
@@ -12,6 +12,70 @@ test_that("invalid token test is specific", {
 })
 
 
+# auth_oauth_token_get() --------------------------------------------------
+
+test_that("can request and cache token if not present", {
+  client <- oauth_client("test", "http://example.org/test")
+  cache <- cache_mem(client)
+
+  token <- oauth_token("123")
+  expect_equal(auth_oauth_token_get(cache, function(...) token), token)
+  expect_equal(cache$get(), token)
+})
+
+test_that("can re-flow to get new token", {
+  client <- oauth_client("test", "http://example.org/test")
+  cache <- cache_mem(client)
+  cache$set(oauth_token("123", expires_in = -60))
+
+  token <- oauth_token("123")
+  expect_equal(auth_oauth_token_get(cache, function(...) token), token)
+  expect_equal(cache$get(), token)
+
+  # and cache is cleared if
+  cache$set(oauth_token("123", expires_in = -60))
+  expect_error(auth_oauth_token_get(cache, function(...) stop("Bad!")))
+  expect_equal(cache$get(), NULL)
+})
+
+test_that("can refresh to get new token", {
+  client <- oauth_client("test", "http://example.org/test")
+  cache <- cache_mem(client)
+  cache$set(oauth_token("123", refresh_token = "456", expires_in = -60))
+
+  local_mocked_bindings(
+    oauth_client_get_token = function(...) oauth_token("789")
+  )
+
+  new_token <- oauth_token("789", refresh_token = "456")
+  expect_equal(auth_oauth_token_get(cache, function(...) NULL), new_token)
+  expect_equal(cache$get(), new_token)
+})
+
+test_that("can reflow if refresh fails", {
+  client <- oauth_client("test", "http://example.org/test")
+  cache <- cache_mem(client)
+  cache$set(oauth_token("123", refresh_token = "456", expires_in = -60))
+
+  local_mocked_bindings(
+    oauth_client_get_token = function(...) oauth_flow_abort("Nope")
+  )
+
+  token <- oauth_token("789")
+  expect_equal(auth_oauth_token_get(cache, function(...) token), token)
+  expect_equal(cache$get(), token)
+})
+
+test_that("can retrieve non-expired token from cache", {
+  client <- oauth_client("test", "http://example.org/test")
+  cache <- cache_mem(client)
+
+  token <- oauth_token("123")
+  cache$set(token)
+  expect_equal(auth_oauth_token_get(cache, oauth_flow_refresh), token)
+})
+
+
 # Cache -------------------------------------------------------------------
 
 test_that("can store in memory", {